requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
PyJWT==2.10.1
bcrypt==4.2.1
python-dotenv==1.0.1
email-validator==2.2.0
